Added smaller development environments
It is now possible to work at a city level directly using the
CityLib.fla. It still publishes to a SWC but you can preview is as a SWF
to see your changes before testing in the main game. I'm going to add a
UnitLib SWC and another city so two cities can be worked on at the same
time while picking up each others code.