Unit SWC added
Party as a test and partly to add the ground work with adding units. I
had to link the unit SWC to the CityLIb so units could be added. Trying
to do it in a way that every type of unit doesn't have to be declared by
a city, so you could add your own types of units dynamically. Needs more
work and maybe some thinking wine, oh I mean music ... nah wine.